Construct iEDGE input object

Construct iEDGE input object

construct_iEDGE(cn, gep, cisgenes, cn.fdat, gep.fdat, cn.pdat = NA,
  gep.pdat = NA)

Arguments

cn
matrix of epi-DNA alterations (m rows of alterations by n samples)

gep
matrix of gene expression (g rows of genes by n samples)

cisgenes
list of cis genes corresponding to each row in cn

cn.fdat
data frame of row annotations for cn, must have column of row identifiers, and optionally column of direction identifiers if one-sided DE tests are applied

gep.fdat
data frame of row annotaions for gep, must have column of row identifiers

cn.pdat
optional data frame of column annotations for cn

gep.pdat
optional data frame of column annotations for gep