Main wrapper for iEDGE, assumes input data is constructed by construct_iEDGE

Main wrapper for iEDGE, assumes input data is constructed by construct_iEDGE

run_iEDGE(dat, header, outdir, gs.file = NA, gepid = "SYMBOL",
  cnid = "Unique.Name", cndesc = "Descriptor",
  cndir = "alteration_direction", fdr.cis.cutoff = 0.25,
  fdr.trans.cutoff = 0.05, fc.cis = NA, fc.trans = NA, min.drawsize = 3,
  onesided.cis = TRUE, onesided.trans = FALSE, uptest = "Amplification",
  downtest = "Deletion", min.group = 2, prune.col = "pvalue",
  prune.thres = 0.05, hyperthres = 0.25, cis.boxplot = TRUE,
  trans.boxplot = TRUE, enrich.heatmap = TRUE, bipartite = TRUE,
  html = TRUE, jsdir = NA, numcores = NA, cache = list(DE = NULL,
  prunning = NULL, ui = NULL), includeheatmap = TRUE)

Arguments

dat
the input iEDGE object, see construct_iEDGE for specification

header
header string for result file names

outdir
output directory

gs.file
default = NA, vector of characters giving full path of gmt files for enrichment analysis

gepid
default = "SYMBOL", colname in fData(dat$gep) indicating unique gene ids to display

cnid
default = "Unique.Name", colname in fData(dat$cn) indicating unique alterations to display

cndesc
default = "Descriptor", column in fData(dat$cn), can be NA if not available

cndir
default = "alteration_direction", column in fData(dat$cn) indicating directions of one sided differential expression, use NA if onesided.cis = FALSE and onesided.trans = FALSE

fdr.cis.cutoff
default = 0.25, fdr cis cutoff

fdr.trans.cutoff
default = 0.05, fdr trans cutoff

fc.cis
default = NA, fold change cutoff cis

fc.trans
default = NA, fold change cutoff trans

min.drawsize
default = 3, min drawsize for geneset enrichment

onesided.cis
default = TRUE, one sided cis differential expression

onesided.trans
default = FALSE, one sided trans differential expression

uptest
default = "Amplification", alterations for which upregulation in alteration is desired, value must be on fData(dat$cn)[, cndir], must specify if onesided.cis =TRUE

downtest
= "Deletion", alterations for which downregulation in alteration is desired

min.group
default = 2, mininum sample size in each group

prune.col
default = "pvalue", column name to indicate significance of sobel test

prune.thres
default = 0.05, significance cutoff of sobel test

hyperthres
default = 0.25, threshold for hyperEnrichment (fdr cutoff)

cis.boxplot
default = TRUE, display boxplot for cis genes

trans.boxplot
default = TRUE, display boxplot for trans genes

enrich.heatmap
= TRUE, display heatmap for pathway enrichment results

bipartite
default = TRUE, do bipartite graph/pruning

html
default = TRUE, do html report

jsdir
default = NA, default directory of iEDGE js files

numcores
default = NA (non-parallel), if specified uses mclapply with specified mc.cores

cache
default = list(DE = NULL, prunning = NULL, ui = NULL), optional, cached result of previous run_iEDGE

includeheatmap
= TRUE, optional, include heatmap for pathway enrichment in UI